[Fix][Ian] change Cart component to dynamically render on the client
